requirements.txt
# requirements.txt — Reviewed and corrected

# ── Core Flask ─────────────────────────────────────────────────────────────────
Flask==3.0.0
Werkzeug==3.0.1

# ── Database ───────────────────────────────────────────────────────────────────
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
# psycopg2-binary==2.9.9   ← REMOVE: you use SQLite, not PostgreSQL.
#                             On some Windows setups this fails to compile.
#                             Add back only if you migrate to PostgreSQL later.

# ── Firebase ───────────────────────────────────────────────────────────────────
firebase-admin==6.3.0
pyrebase4==4.7.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0

# ── Authentication & Security ──────────────────────────────────────────────────
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
bcrypt==4.1.2
PyJWT==2.8.0

# ── Rate Limiting ──────────────────────────────────────────────────────────────
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
limits==3.7.0

# ── Email ──────────────────────────────────────────────────────────────────────
Flask-Mail==0.9.1

# ── QR Code ────────────────────────────────────────────────────────────────────
qrcode[pil]==7.4.2
Pillow==10.2.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.29.0
validators==0.22.0

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
coverage==7.4.0

# ── Production Server ──────────────────────────────────────────────────────────
gunicorn==21.2.0

# ── Scheduler ──────────────────────────────────────────────────────────────────
Flask-APScheduler==1.13.1
APScheduler==3.11.2    # ← Pin to exact version (>=3.6.3,<4 is ambiguous in pip freeze)
